added tslint + prettierrc config all the classes are now in their own file. ready to publish
