+++ .npmignore
@@ -1,1 +1,4 @@
-dist/
+src
+tsconfig.json
+tslint.json
+.prettierrc
