+++ package.json
@@ -1,12 +1,14 @@
 {
   "name": "angles-javascript-client",
-  "version": "1.0.0-BETA1",
-  "description": "In development",
+  "version": "1.0.0-BETA2",
+  "description": "This is the javascript client for the Angles Dashboard. It allows you to store your test results.",
   "main": "index.js",
   "types": "index.d.ts",
   "scripts": {
-    "build": "tsc",
-    "start": "node dist/index.js"
+    "build": "rm -rf dist/ && tsc",
+    "start": "node dist/index.js",
+    "format": "prettier --write \"src/**/*.ts\" \"dist/**/*.js\"",
+    "lint": "tslint -p tsconfig.json"
   },
   "repository": {
     "type": "git",
@@ -32,11 +34,12 @@
   },
   "devDependencies": {
     "@types/node": "^14.14.37",
-    "@typescript-eslint/eslint-plugin": "^4.20.0",
-    "@typescript-eslint/parser": "^4.20.0",
-    "eslint": "^7.23.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-plugin-import": "^2.22.1",
+    "prettier": "^2.2.1",
+    "tslint": "^6.1.3",
+    "tslint-config-prettier": "^1.18.0",
     "typescript": "^4.2.3"
-  }
+  },
+  "files": [
+    "dist/**/*"
+  ]
 }
