+++ src/index.ts
@@ -1,11 +1,10 @@
-import { AnglesReporterClass } from './lib/AnglesReporter'
-import { Team, Environment , Action, Step, StepStates } from './lib/models/Types'
-import { ScreenshotPlatform, StoreScreenshot } from './lib/models/RequestTypes';
-import { Artifact } from '../dist/lib/types';
+import { AnglesReporterClass } from './lib/AnglesReporter';
+// import { Artifact } from '../dist/lib/types';
+// import { ScreenshotPlatform } from './lib/models/requests/ScreenshotPlatform';
 
-const reporter = AnglesReporterClass.getInstance();
+// const reporter = AnglesReporterClass.getInstance();
 
-const runTest = async () => {
+// const runTest = async () => {
   // const teams = await reporter.teams.getTeams();
   // teams.forEach((team:Team) => {
   //   console.log(`${team.name} - ${team._id}`);
@@ -20,40 +19,47 @@ const runTest = async () => {
   //   console.log(`Environment: ${environment.name}`);
   // })
 
-  try {
-    // start build
-    await reporter.startBuild("angles-javascript-client", "qa", "qa", "regression");
+  // try {
+  //   // start build
+  //   await reporter.startBuild('angles-javascript-client', 'qa', 'qa', 'regression');
+  //
+  //   // add artifacts
+  //   const artifact = new Artifact();
+  //   artifact.artifactId = 'angles-ui';
+  //   artifact.groupId = 'anglesHQ';
+  //   artifact.version = '1.0.0';
+  //   const artifactArray: Artifact[] = [];
+  //   artifactArray.push(artifact);
+  //   await reporter.addArtifacts(artifactArray);
+  //
+  //   // start test
+  //   reporter.startTest('test1', 'suite1');
+  //
+  //   // add action
+  //   reporter.addAction('My first action');
+  //
+  //   // start reporting examples
+  //   const platform = new ScreenshotPlatform();
+  //   platform.platformName = 'Android';
+  //   platform.browserName = 'Chrome';
+  //   const screenshot = await reporter.saveScreenshotWithPlatform(
+  //     '/Users/sergio.barros/Desktop/image.png',
+  //     'view_1',
+  //     platform,
+  //   );
+  //   reporter.infoWithScreenshot('Took Screenshot', screenshot._id);
+  //   reporter.pass('Assertion', 'true', 'true', 'Just doing an assertion');
+  //
+  //   // store the test
+  //   await reporter.saveTest();
+  // } catch (error) {
+  //   // tslint:disable-next-line:no-console
+  //   console.log(error);
+  // }
+// };
 
-    // add artifacts
-    const artifact = new Artifact();
-    artifact.artifactId = 'angles-ui';
-    artifact.groupId = 'anglesHQ';
-    artifact.version = '1.0.0';
-    const artifactArray: Artifact[] = [];
-    artifactArray.push(artifact);
-    await reporter.addArtifacts(artifactArray);
+// runTest();
 
-    // start test
-    reporter.startTest("test1", "suite1");
-
-    // add action
-    reporter.addAction("My first action");
-
-    // start reporting examples
-    const platform = new ScreenshotPlatform();
-    platform.platformName = "Android";
-    platform.browserName = "Chrome";
-    const screenshot = await reporter.saveScreenshotWithPlatform("/Users/sergio.barros/Desktop/image.png", "view_1", platform);
-    reporter.infoWithScreenshot("Took Screenshot", screenshot._id);
-    reporter.pass("Assertion", "true", "true", "Just doing an assertion");
-
-    // store the test
-    await reporter.saveTest();
-
-  } catch(error) {
-    console.log(error);
-  }
-
+export {
+  AnglesReporterClass
 }
-
-runTest();
