+++ src/lib/AnglesReporter.ts
@@ -1,26 +1,28 @@
-import axios, { AxiosInstance} from "axios";
+import axios, { AxiosInstance } from 'axios';
 import { TeamRequests } from './requests/TeamRequests';
 import { EnvironmentRequests } from './requests/EnvironmentRequests';
 import { BuildRequests } from './requests/BuildRequests';
 import { ExecutionRequests } from './requests/ExecutionRequests';
 import { ScreenshotRequests } from './requests/ScreenshotRequests';
-import {
-  CreateBuild,
-  CreateExecution,
-  ScreenshotPlatform,
-  StoreScreenshot
-} from './models/RequestTypes';
-import { Action, Artifact, Build, Execution, Screenshot, Step, StepStates } from './models/Types';
+import { Build } from './models/Build';
+import { CreateExecution } from './models/requests/CreateExecution';
+import { Action } from './models/Action';
+import { CreateBuild } from './models/requests/CreateBuild';
+import { Artifact } from './models/Artifact';
+import { Screenshot } from './models/Screenshot';
+import { StoreScreenshot } from './models/requests/StoreScreenshot';
+import { StepStates } from './models/enum/StepStates';
+import { Step } from './models/Step';
+import { ScreenshotPlatform } from './models/requests/ScreenshotPlatform';
 
 export class AnglesReporterClass {
-
-  private static _instance:AnglesReporterClass = new AnglesReporterClass();
+  private static _instance: AnglesReporterClass = new AnglesReporterClass();
   protected axiosInstance: AxiosInstance;
-  public teams:TeamRequests;
-  public environments:EnvironmentRequests;
-  public builds:BuildRequests;
-  public executions:ExecutionRequests;
-  public screenshots:ScreenshotRequests;
+  public teams: TeamRequests;
+  public environments: EnvironmentRequests;
+  public builds: BuildRequests;
+  public executions: ExecutionRequests;
+  public screenshots: ScreenshotRequests;
 
   private currentBuild: Build;
   private currentExecution: CreateExecution;
@@ -34,14 +36,14 @@ export class AnglesReporterClass {
     const apiConfig = {
       returnRejectedPromiseOnError: true,
       timeout: 10000,
-      baseURL: "http://127.0.0.1:3000/rest/api/v1.0/",
+      baseURL: 'http://127.0.0.1:3000/rest/api/v1.0/',
       headers: {
         common: {
-          "Content-Type": "application/json",
-          Accept: "application/json",
+          'Content-Type': 'application/json',
+          Accept: 'application/json',
         },
       },
-    }
+    };
     this.axiosInstance = axios.create(apiConfig);
     this.teams = new TeamRequests(this.axiosInstance);
     this.environments = new EnvironmentRequests(this.axiosInstance);
@@ -50,17 +52,12 @@ export class AnglesReporterClass {
     this.screenshots = new ScreenshotRequests(this.axiosInstance);
   }
 
-  public static getInstance():AnglesReporterClass {
+  public static getInstance(): AnglesReporterClass {
     return AnglesReporterClass._instance;
   }
 
-  public async startBuild(name:string, team:string, environment:string, component: string): Promise<Build> {
-    const createBuildRequest = new CreateBuild(
-      name,
-      environment,
-      team,
-      component
-    );
+  public async startBuild(name: string, team: string, environment: string, component: string): Promise<Build> {
+    const createBuildRequest = new CreateBuild(name, environment, team, component);
     this.currentBuild = await this.builds.createBuild(createBuildRequest);
     return this.currentBuild;
   }
@@ -69,7 +66,7 @@ export class AnglesReporterClass {
     return await this.builds.addArtifacts(this.currentBuild._id, artifacts);
   }
 
-  public startTest(title:string, suite: string):void {
+  public startTest(title: string, suite: string): void {
     this.currentExecution = new CreateExecution();
     this.currentExecution.title = title;
     this.currentExecution.suite = suite;
@@ -79,14 +76,18 @@ export class AnglesReporterClass {
 
   public async saveTest() {
     this.currentAction = undefined;
-    return await this.executions.saveExecution(this.currentExecution)
+    return await this.executions.saveExecution(this.currentExecution);
   }
 
-  public async saveScreenshot(filePath:string, view:string): Promise<Screenshot>  {
+  public async saveScreenshot(filePath: string, view: string): Promise<Screenshot> {
     return await this.saveScreenshotWithPlatform(filePath, view, undefined);
   }
 
-  public async saveScreenshotWithPlatform(filePath:string, view:string, platform:ScreenshotPlatform): Promise<Screenshot> {
+  public async saveScreenshotWithPlatform(
+    filePath: string,
+    view: string,
+    platform: ScreenshotPlatform,
+  ): Promise<Screenshot> {
     const storeScreenshot = new StoreScreenshot();
     storeScreenshot.buildId = this.currentBuild._id;
     storeScreenshot.filePath = filePath;
@@ -99,7 +100,7 @@ export class AnglesReporterClass {
     }
   }
 
-  public addAction(name:string) {
+  public addAction(name: string) {
     this.currentAction = new Action();
     this.currentAction.name = name;
     this.currentAction.start = new Date();
@@ -107,39 +108,46 @@ export class AnglesReporterClass {
     this.currentExecution.actions.push(this.currentAction);
   }
 
-  public info(info:string) {
-    this.addStep("INFO",undefined,undefined,info, StepStates.INFO, undefined);
+  public info(info: string) {
+    this.addStep('INFO', undefined, undefined, info, StepStates.INFO, undefined);
   }
 
-  public infoWithScreenshot(info:string, screenshotId: string) {
-    this.addStep("INFO",undefined,undefined,info, StepStates.INFO, screenshotId);
+  public infoWithScreenshot(info: string, screenshotId: string) {
+    this.addStep('INFO', undefined, undefined, info, StepStates.INFO, screenshotId);
   }
 
-  public error(error:string) {
-    this.addStep("ERROR",undefined,undefined,error, StepStates.ERROR, undefined);
+  public error(error: string) {
+    this.addStep('ERROR', undefined, undefined, error, StepStates.ERROR, undefined);
   }
 
   public errorWithScreenshot(error: string, screenshotId: string) {
-    this.addStep("ERROR",undefined,undefined,error, StepStates.ERROR, screenshotId);
+    this.addStep('ERROR', undefined, undefined, error, StepStates.ERROR, screenshotId);
   }
 
-  public pass(name:string, expected:string, actual:string, info:string): void {
-    this.addStep(name,expected,actual,info,StepStates.PASS, undefined);
+  public pass(name: string, expected: string, actual: string, info: string): void {
+    this.addStep(name, expected, actual, info, StepStates.PASS, undefined);
   }
 
-  public passWithScreenshot(name:string, expected:string, actual:string, info:string, screenshot: string): void {
-    this.addStep(name,expected,actual,info,StepStates.PASS, screenshot);
+  public passWithScreenshot(name: string, expected: string, actual: string, info: string, screenshot: string): void {
+    this.addStep(name, expected, actual, info, StepStates.PASS, screenshot);
   }
 
-  public fail(name:string, expected:string, actual:string, info:string) {
-    this.addStep(name,expected,actual,info,StepStates.FAIL, undefined);
+  public fail(name: string, expected: string, actual: string, info: string) {
+    this.addStep(name, expected, actual, info, StepStates.FAIL, undefined);
   }
 
-  public failWithScreenshot(name:string, expected:string, actual:string, info:string, screenshotId: string) {
-    this.addStep(name,expected,actual,info,StepStates.FAIL, screenshotId);
+  public failWithScreenshot(name: string, expected: string, actual: string, info: string, screenshotId: string) {
+    this.addStep(name, expected, actual, info, StepStates.FAIL, screenshotId);
   }
 
-  public addStep(name: string, expected: string, actual:string, info: string, status: StepStates, screenshot:string):void {
+  public addStep(
+    name: string,
+    expected: string,
+    actual: string,
+    info: string,
+    status: StepStates,
+    screenshot: string,
+  ): void {
     if (this.currentAction === undefined) {
       this.addAction('test-details');
     }
@@ -153,5 +161,4 @@ export class AnglesReporterClass {
     step.screenshot = screenshot;
     this.currentAction.steps.push(step);
   }
-
 }
