+++ src/lib/models/Artifact.ts
@@ -1,0 +1,5 @@
+export class Artifact {
+  groupId: string;
+  artifactId: string;
+  version: string;
+}
