+++ src/lib/models/Environment.ts
@@ -1,0 +1,4 @@
+export class Environment {
+  _id: number;
+  name: string;
+}
