+++ src/lib/models/Platform.ts
@@ -1,0 +1,12 @@
+export class Platform {
+  _id: string;
+  platformName: string;
+  platformVersion: string;
+  browserName: string;
+  browserVersion: string;
+  deviceName: string;
+  userAgent: string;
+  screenHeight: number;
+  screenWidth: number;
+  pixelRatio: number;
+}
