+++ src/lib/models/Team.ts
@@ -1,0 +1,10 @@
+export class Team {
+  _id: number;
+  name: string;
+  components: [
+    {
+      name: string;
+      _id: string;
+    },
+  ];
+}
