+++ src/lib/models/enum/ExecutionStates.ts
@@ -1,0 +1,6 @@
+export enum ExecutionStates {
+  SKIPPED = 'SKIPPED',
+  PASS = 'PASS',
+  ERROR = 'ERROR',
+  FAIL = 'FAIL',
+}
