+++ src/lib/models/enum/StepStates.ts
@@ -1,0 +1,7 @@
+export enum StepStates {
+  INFO = 'INFO',
+  DEBUG = 'DEBUG',
+  PASS = 'PASS',
+  ERROR = 'ERROR',
+  FAIL = 'FAIL',
+}
