+++ src/lib/models/requests/ScreenshotPlatform.ts
@@ -1,0 +1,7 @@
+export class ScreenshotPlatform {
+  platformName: string;
+  platformVersion: string;
+  browserName: string;
+  browserVersion: string;
+  deviceName: string;
+}
