+++ src/lib/models/requests/StoreScreenshot.ts
@@ -1,0 +1,6 @@
+export class StoreScreenshot {
+  buildId: string;
+  view: string;
+  timestamp: Date;
+  filePath: string;
+}
