+++ src/lib/requests/BaseRequests.ts
@@ -1,9 +1,7 @@
 import { AxiosResponse } from 'axios';
 
 export class BaseRequests {
-
   protected success<T>(response: AxiosResponse<T>): T {
     return response.data;
   }
-
 }
