+++ src/lib/requests/BuildRequests.ts
@@ -1,18 +1,18 @@
 import { AxiosResponse, AxiosInstance } from 'axios';
 import { BaseRequests } from './BaseRequests';
-import { Artifact, Build } from '../models/Types';
-import { CreateBuild } from '../models/RequestTypes';
+import { CreateBuild } from '../models/requests/CreateBuild';
+import { Build } from '../models/Build';
+import { Artifact } from '../models/Artifact';
 
 export class BuildRequests extends BaseRequests {
-
   private axios: AxiosInstance;
 
-  public constructor (axiosInstance: AxiosInstance) {
+  public constructor(axiosInstance: AxiosInstance) {
     super();
     this.axios = axiosInstance;
   }
 
-  public async getBuild(buildId:string): Promise<Build> {
+  public async getBuild(buildId: string): Promise<Build> {
     const response: AxiosResponse<Build> = await this.axios.get<Build>(`build/${buildId}`);
     return this.success(response);
   }
@@ -28,13 +28,17 @@ export class BuildRequests extends BaseRequests {
   }
 
   public async addArtifacts(buildId: string, artifacts: Artifact[]): Promise<Build> {
-    const response: AxiosResponse<Build> = await this.axios.put<Build>(`build/${buildId}/artifacts`, {
-      artifacts: artifacts
-    },  {
-      headers: {
-        'Content-Type': 'application/json',
-      }
-    });
+    const response: AxiosResponse<Build> = await this.axios.put<Build>(
+      `build/${buildId}/artifacts`,
+      {
+        artifacts,
+      },
+      {
+        headers: {
+          'Content-Type': 'application/json',
+        },
+      },
+    );
     return this.success(response);
   }
 }
