+++ src/lib/requests/ExecutionRequests.ts
@@ -1,18 +1,17 @@
 import { AxiosResponse, AxiosInstance } from 'axios';
 import { BaseRequests } from './BaseRequests';
-import { Execution } from '../models/Types';
-import { CreateExecution } from '../models/RequestTypes';
+import { Execution } from '../models/Execution';
+import { CreateExecution } from '../models/requests/CreateExecution';
 
 export class ExecutionRequests extends BaseRequests {
-
   private axios: AxiosInstance;
 
-  public constructor (axiosInstance: AxiosInstance) {
+  public constructor(axiosInstance: AxiosInstance) {
     super();
     this.axios = axiosInstance;
   }
 
-  public async getExecution(executionId:string): Promise<Execution> {
+  public async getExecution(executionId: string): Promise<Execution> {
     const response: AxiosResponse<Execution> = await this.axios.get<Execution>(`execution/${executionId}`);
     return this.success(response);
   }
@@ -21,5 +20,4 @@ export class ExecutionRequests extends BaseRequests {
     const response: AxiosResponse<Execution> = await this.axios.post<Execution>(`execution/`, saveExecutionRequest);
     return this.success(response);
   }
-
 }
