+++ tsconfig.json
@@ -1,4 +1,3 @@
-
 {
   "compilerOptions": {
     "strict": true,
